Allow Kafka configuration with -X
This commit gives the ability to set global configuration
properties for Kafka by use of the -X option to a compiled
Souffle program.

For example, if we have a program called `my_program`, then
running `my_program -Xmetadata.broker.list=localhost:9092`
will set the Kafka broker to `localhost:9092`.

For a full list of available configuration properties, see
the "Global configuration properties" section of the document at
https://github.com/edenhill/librdkafka/blob/master/CONFIGURATION.md.